chore: move utils
